// Kulthia Dasboard/src/app/core/models/sub-item.model.ts
export interface SubItemModel {
    id: Number;
    SlNo: any;
 
    ItemName: string;
    HSNCode: string;
    SubItem: string;
    BaseUnit: string;
    PurRt:Number;
    JobDiscount: Number;
    GoldItem: string;
    Purity: Number;
    OpRt: Number;
    OpPcs: Number;
    PureQty: Number;
    OpAmt: Number;
    ItemCode: Number;
    SItemCode: Number;
}